New module, OdieRun, to produce Garfield++ gas files.
OdieRun is based on PyBoltzRun and provides extra functionality to
produce Garfield++ formatted gas files, automate generating gas
properties on a grid, and read simulation settings from JSON files.

OdieRun should be able to be used in place of PyBoltzRun as it shares
the same basic interface, but has been extended to include new features.

New example script to show how to use Odie to generate a Garfield++ gas
file from a grid of EFields, and how to read simulation settings from a
JSON file.

Included examples of a JSON input file and output gas file.

Updated README.md files to include descriptions of the new Odie module
and files.