Update gas file output to use alpha/eta SST values
Updated OdieRun to use the alpha and eta SST calculation values in the
gas file when the steady state threshold is crossed.

Previously the gas file would always store the first calculation of
alpha and eta. These values can be quite different from the later
calculations of alpha and eta from the SST, PT, or TOF simulations.

MagBoltz by default reports the SST calculations of alpha and eta if the
steady state threshold has been crossed. OdieRun now does the same in
the gas file.

The Boltz class was modified to include a boolean flag to indicate if
the steady state threshold has been crossed.

NB: If the steady state threshold was crossed, but the SST simulation
was not run, the flag is set to False as the alpha and eta SST values
are not calculated.