Added alternative names for gases
The types of gases in a given mix are specified by strings in the
settings dictionary. This commit adds alternative names/abbriviations
for most gases for a more user friendly experience. Additionally the gas
strings are now converted to upper case before comparison which makes
them case-insensitive to the user.

For example, PyBoltz now accepts N2, n2, or Nitrogen as a valid string for
nitrogen gas instead of only NITROGEN.

The ListGases() function now prints all accepted strings for each gas.

Edited the printing of the settings dictionary to use the JSON pretty
printing functionality.